requirements.txt
chainlit==2.9.4
fastapi==0.127.0
httpx==0.28.1
llama_index==0.14.10
pydantic==2.12.5
pydantic_settings==2.12.0
python-dotenv==1.2.1
python_docx==1.2.0
qdrant_client==1.16.2
tqdm==4.66.2
uvicorn==0.40.0
# Additional llama-index plugins (cài những plugin bổ sung cho llama-index, có thể cài sau khi cài  những thư viện trên tránh xung đột)
llama-index-postprocessor-sbert-rerank==0.4.2
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-index-retrievers-bm2
